Renamed entrypoint classes + fixed scalafmt warning
