+++ nsdb-cluster/src/main/scala/io/radicalbit/nsdb/cluster/NsdbCluster.scala
@@ -23,4 +23,4 @@ import io.radicalbit.nsdb.web.Web
 /**
   * Run a concrete Nsdb cluster node according to the configuration provided in `confDir` folder or into the classpath
   */
-object Cluster extends App with ProductionCluster with Web with NsdbSecurity with NsdbConfig with LazyLogging
+object NsdbCluster extends App with ProductionCluster with Web with NsdbSecurity with NsdbConfig with LazyLogging
